% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters.R
\name{add_cluster_circles}
\alias{add_cluster_circles}
\title{Add Cluster Circles to Map}
\usage{
add_cluster_circles(
  map,
  cluster_data,
  color_palette,
  opacity,
  use_radius = TRUE,
  is_simple = FALSE,
  ...
)
}
\description{
Add Cluster Circles to Map
}
\keyword{internal}
