% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters.R
\name{add_location_points}
\alias{add_location_points}
\title{Add Location Points to Map}
\usage{
add_location_points(map, location_data, color, coords)
}
\description{
Add Location Points to Map
}
\keyword{internal}
