% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_options.R
\name{apply_user_overrides}
\alias{apply_user_overrides}
\title{Apply User Overrides}
\usage{
apply_user_overrides(opts, user_opts)
}
\arguments{
\item{opts}{Current options list}

\item{user_opts}{Named list of user overrides}
}
\value{
Merged options list
}
\description{
Merge user options with current options (user takes precedence)
}
\keyword{internal}
