% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{as_satscan_control}
\alias{as_satscan_control}
\title{Coerce Data to SaTScan Control Table}
\usage{
as_satscan_control(
  data,
  loc_id,
  controls,
  time = NULL,
  covars = NULL,
  spec = list()
)
}
\arguments{
\item{data}{A data frame containing control data.}

\item{loc_id}{Column name for location ID.}

\item{controls}{Column name for control counts.}

\item{time}{Optional column name for time.}

\item{covars}{Optional vector of column names for covariates.}

\item{spec}{Additional specifications.}
}
\value{
An \code{ss_tbl} object of type "ctl".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "ctl" (Control File) for Bernoulli models.
}
