% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{as_satscan_grid}
\alias{as_satscan_grid}
\title{Coerce Data to SaTScan Grid Table}
\usage{
as_satscan_grid(
  data,
  coord1,
  coord2,
  z = NULL,
  earliest_start = NULL,
  latest_start = NULL,
  earliest_end = NULL,
  latest_end = NULL,
  spec = list(grid_variant = "basic")
)
}
\arguments{
\item{data}{A data frame containing grid locations.}

\item{coord1}{Column name for the first coordinate.}

\item{coord2}{Column name for the second coordinate.}

\item{z}{Optional column name for Z-coordinate.}

\item{earliest_start}{Optional column for earliest start time constraint.}

\item{latest_start}{Optional column for latest start time constraint.}

\item{earliest_end}{Optional column for earliest end time constraint.}

\item{latest_end}{Optional column for latest end time constraint.}

\item{spec}{Additional specifications (e.g. \code{grid_variant}).}
}
\value{
An \code{ss_tbl} object of type "grd".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "grd" (Grid File).
}
