% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{as_satscan_meta_locations}
\alias{as_satscan_meta_locations}
\title{Coerce Data to SaTScan Meta Locations}
\usage{
as_satscan_meta_locations(data, meta_loc_id, member_cols)
}
\arguments{
\item{data}{A data frame containing meta-location definitions.}

\item{meta_loc_id}{Column name for the meta-location ID.}

\item{member_cols}{Character vector of column names representing member locations.}
}
\value{
An \code{ss_tbl} object of type "met".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "met" (Meta Population File).
}
