% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{as_satscan_neighbors}
\alias{as_satscan_neighbors}
\title{Coerce Data to SaTScan Neighbors Table}
\usage{
as_satscan_neighbors(data, loc_id, neighbor_cols, spec = list())
}
\arguments{
\item{data}{A data frame containing neighbor lists.}

\item{loc_id}{Column name for the location ID.}

\item{neighbor_cols}{Character vector of column names representing neighbors.}

\item{spec}{Additional specifications.}
}
\value{
An \code{ss_tbl} object of type "nbr".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "nbr" (Neighbors File).
}
