% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{as_satscan_network}
\alias{as_satscan_network}
\title{Coerce Data to SaTScan Network Table}
\usage{
as_satscan_network(
  data,
  loc_id,
  neighbor_id,
  distance = NULL,
  spec = list(distance_units = NULL)
)
}
\arguments{
\item{data}{A data frame containing network edges.}

\item{loc_id}{Column name for the source location ID.}

\item{neighbor_id}{Column name for the connected neighbor ID.}

\item{distance}{Optional column name for distance/weight for the connection.}

\item{spec}{Additional specifications (e.g. \code{distance_units}).}
}
\value{
An \code{ss_tbl} object of type "nwk".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "nwk" (Network File).
}
