% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{as_satscan_population}
\alias{as_satscan_population}
\title{Coerce Data to SaTScan Population Table}
\usage{
as_satscan_population(
  data,
  loc_id,
  time,
  population,
  covars = NULL,
  time_precision = NULL
)
}
\arguments{
\item{data}{A data frame containing population data.}

\item{loc_id}{Column name for location ID.}

\item{time}{Column name for time (census year/time).}

\item{population}{Column name for population count.}

\item{covars}{Optional vector of column names for covariates.}

\item{time_precision}{User-specified time precision (e.g. "year" to format Dates as YYYY).}
}
\value{
An \code{ss_tbl} object of type "pop".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "pop" (Population File).
}
