% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{attach_analysis_params}
\alias{attach_analysis_params}
\title{Store Analysis Parameters in Result}
\usage{
attach_analysis_params(result, params)
}
\arguments{
\item{result}{A \code{satscan_result} object.}

\item{params}{List of analysis parameters.}
}
\value{
The result object with parameters attached.
}
\description{
Adds analysis parameters as attributes to the satscan_result object.
This is intended to be called internally by satscanr().
}
\keyword{internal}
