% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_build_location_code.R
\name{build_location_code}
\alias{build_location_code}
\title{Build Location Code from Administrative Components}
\usage{
build_location_code(
  data,
  prefix_col,
  middle_col,
  suffix_col,
  output_col = "location_id",
  pad_width = c(2, 2, 2)
)
}
\arguments{
\item{data}{Data frame containing administrative code columns}

\item{prefix_col}{Character. Name of column containing prefix/province code}

\item{middle_col}{Character. Name of column containing middle/district code}

\item{suffix_col}{Character. Name of column containing suffix/subdistrict code}

\item{output_col}{Character. Name of output column (default: "location_id")}

\item{pad_width}{Integer vector. Number of digits for each component (default: c(2,2,2))}
}
\value{
Data frame with new location_id column
}
\description{
Constructs a standardized location identifier by concatenating administrative
unit codes (e.g., province, district, subdistrict).
}
\examples{
data <- data.frame(chw_code = 90, amp_code = 1, tmb_code = 5)
build_location_code(data, "chw_code", "amp_code", "tmb_code")
# Returns data with location_id = "900105"

}
