% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_analysis.R
\name{calculate_kde}
\alias{calculate_kde}
\title{Spatial Odds Analysis Functions}
\usage{
calculate_kde(points, study_area, bandwidth, resolution = 100)
}
\arguments{
\item{points}{sf object of case locations}

\item{study_area}{sf object of the study area boundary}

\item{bandwidth}{Bandwidth (sigma) for the Gaussian kernel (in map units, e.g., meters)}

\item{resolution}{Resolution of the output raster (in map units)}
}
\value{
A terra raster object representing the density
}
\description{
This script contains functions for performing Kernel Density Estimation (KDE)
and calculating Spatial Odds Ratios (SOR) for disease clustering analysis.
}
