% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_analysis.R
\name{calculate_spatial_odds}
\alias{calculate_spatial_odds}
\title{Calculate Spatial Odds Ratio}
\usage{
calculate_spatial_odds(case_kde, pop_raster)
}
\arguments{
\item{case_kde}{Raster of case density}

\item{pop_raster}{Raster of population density (control)}
}
\value{
Raster of Spatial Odds Ratio (Case / Pop)
}
\description{
Calculate Spatial Odds Ratio
}
