% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_check.R
\name{check_installation}
\alias{check_installation}
\title{Check Package Installation}
\usage{
check_installation()
}
\value{
Logical. TRUE if all packages are found.
}
\description{
Verifies that all required packages for epidscan are installed.
}
