% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pop_model.R
\name{compare_population_methods}
\alias{compare_population_methods}
\title{Compare Population Methods}
\usage{
compare_population_methods(pop_raw, pop_expected)
}
\arguments{
\item{pop_raw}{Data frame with raw population}

\item{pop_expected}{Data frame with expected cases}
}
\value{
Data frame with comparison
}
\description{
Compares raw population vs expected cases for each location.
}
