% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{compute_cluster_summary}
\alias{compute_cluster_summary}
\title{Compute Cluster Summary Statistics}
\usage{
compute_cluster_summary(clusters)
}
\arguments{
\item{clusters}{Cluster data frame}
}
\value{
Data frame with summary stats
}
\description{
Compute Cluster Summary Statistics
}
\keyword{internal}
