% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{compute_location_summary}
\alias{compute_location_summary}
\title{Compute Location Summary Statistics}
\usage{
compute_location_summary(locations)
}
\arguments{
\item{locations}{Location data frame}
}
\value{
List with summary info
}
\description{
Compute Location Summary Statistics
}
\keyword{internal}
