% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_viz.R
\name{create_cluster_boundaries}
\alias{create_cluster_boundaries}
\title{Create Cluster Boundaries}
\usage{
create_cluster_boundaries(map_data)
}
\arguments{
\item{map_data}{sf object from prepare_cluster_map_data()}
}
\value{
sf object with cluster boundaries
}
\description{
Generates cluster boundary polygons by unioning locations within each cluster.
}
