% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters.R
\name{detect_coordinates}
\alias{detect_coordinates}
\title{Detect Coordinate Columns}
\usage{
detect_coordinates(df)
}
\arguments{
\item{df}{Data frame with location coordinates}
}
\value{
List with lat_col and lon_col
}
\description{
Detect Coordinate Columns
}
\keyword{internal}
