% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_detect_spatial_extent.R
\name{detect_spatial_extent}
\alias{detect_spatial_extent}
\title{Detect Spatial Extent from Case Data}
\usage{
detect_spatial_extent(case_data, location_col = "location_id")
}
\arguments{
\item{case_data}{Case data frame or path to RDS file}

\item{location_col}{Name of location ID column (default: "location_id")}
}
\value{
Character vector of unique province codes
}
\description{
Automatically detects province codes from case data location IDs.
}
\details{
Extracts first 2 digits from location IDs to determine provinces.
Example: "900105" -> "90" (Songkhla province)
}
\examples{
# From data frame
detect_spatial_extent(cases_df)

# From file
detect_spatial_extent("data/cases_prepared.rds")

}
