% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_filter_spatial_extent.R
\name{filter_spatial_extent}
\alias{filter_spatial_extent}
\title{Filter Population by Spatial Extent}
\usage{
filter_spatial_extent(data, province_codes, location_col = "areacode")
}
\arguments{
\item{data}{Population data frame with areacode column}

\item{province_codes}{Character vector of province codes (e.g., c("90", "91"))}

\item{location_col}{Column containing location codes (default: "areacode")}
}
\value{
Filtered data frame
}
\description{
Filters population data to specific province codes.
}
\examples{
filter_spatial_extent(pop_data, c("90")) # Songkhla only
filter_spatial_extent(pop_data, c("90", "91")) # Songkhla + Satun

}
