% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{format_satscan_time}
\alias{format_satscan_time}
\title{Format Time for SaTScan}
\usage{
format_satscan_time(x, time_precision)
}
\arguments{
\item{x}{Vector of times (Date, POSIXt, numeric, character)}

\item{time_precision}{"day", "month", "year", or "generic"}
}
\value{
Character vector formatted for SaTScan
}
\description{
Format Time for SaTScan
}
\keyword{internal}
