% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_results.R
\name{get_macos_satscan_path}
\alias{get_macos_satscan_path}
\title{Get macOS SatScan Path}
\usage{
get_macos_satscan_path(ss_full_path)
}
\arguments{
\item{ss_full_path}{Full path to SatScan executable}
}
\value{
List with ss_location and ss_batch
}
\description{
Handles macOS App Bundle path detection
}
\keyword{internal}
