% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{get_params}
\alias{get_params}
\title{Extract Analysis Parameters from Result}
\usage{
get_params(x)
}
\arguments{
\item{x}{A \code{satscan_result} object.}
}
\value{
List of parameters, or NULL if not stored.
}
\description{
Retrieves stored analysis parameters from a satscan_result object.
}
