% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{get_satscan_metadata}
\alias{get_satscan_metadata}
\title{Extract Human-Readable Metadata}
\usage{
get_satscan_metadata(x)
}
\arguments{
\item{x}{A \code{satscan_result} object.}
}
\value{
List of metadata
}
\description{
Extract Human-Readable Metadata
}
\keyword{internal}
