% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{get_satscan_path}
\alias{get_satscan_path}
\title{Get SatScan Executable Path}
\usage{
get_satscan_path()
}
\value{
Character path or NULL if not set.
}
\description{
Retrieves the configured SatScan path.
}
