% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{get_satscan_version}
\alias{get_satscan_version}
\title{Get SaTScan Version}
\usage{
get_satscan_version()
}
\value{
Character string containing the version (e.g. "10.2.5") or NULL if unable to determine.
}
\description{
Attempts to determine the version of the configured SaTScan executable.
}
