% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{glance.satscan_result}
\alias{glance.satscan_result}
\title{Glance at SaTScan Results}
\usage{
glance.satscan_result(x, ...)
}
\arguments{
\item{x}{A \code{satscan_result} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A one-row data frame with analysis-level statistics:
\describe{
\item{n_clusters}{Number of clusters detected}
\item{n_significant}{Number of significant clusters (p < 0.05)}
\item{n_locations}{Total locations analyzed}
\item{n_in_clusters}{Locations within detected clusters}
\item{min_p_value}{Minimum (best) p-value}
\item{max_relative_risk}{Maximum relative risk}
\item{prop_in_clusters}{Proportion of locations in clusters}
}
}
\description{
Returns a single-row summary of the entire analysis, following the broom
package convention. Useful for comparing multiple analyses.
}
\examples{
\dontrun{
result <- satscanr(cas, geo, pop, ...)
glance(result)
}

}
