% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscanr.R
\name{infer_dates_from_data}
\alias{infer_dates_from_data}
\title{Infer Dates from Data
Infer Dates from Data}
\usage{
infer_dates_from_data(
  current_opts,
  time_values,
  time_precision_char,
  verbose = FALSE
)
}
\arguments{
\item{current_opts}{List of current SaTScan options}

\item{time_precision_char}{Character string: "day", "month", or "year"}

\item{verbose}{Logical, print messages}

\item{cas_data}{Case data frame (must have a 'time' column)}
}
\value{
Named list of inferred dates (StartDate, EndDate) or NULL if nothing inferred.
}
\description{
Helper to infer StartDate and EndDate from the case data if missing from options.
}
\keyword{internal}
