% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_load_from_cache.R
\name{load_from_cache}
\alias{load_from_cache}
\title{Load Population Data from Cache}
\usage{
load_from_cache(years, cache_dir = "data/population_cache")
}
\arguments{
\item{years}{Numeric vector of years to load}

\item{cache_dir}{Cache directory path}
}
\value{
Data frame with combined population data from all years
}
\description{
Loads cached population data for specified years.
}
\examples{
pop_data <- load_from_cache(c(2023, 2025))

}
