% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{make_tidy_df}
\alias{make_tidy_df}
\title{Create Tidy Data Frame}
\usage{
make_tidy_df(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
Tibble if available, otherwise data.frame
}
\description{
Converts to tibble if available, otherwise returns data.frame.
}
\keyword{internal}
