% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_table.R
\name{new_satscan_table}
\alias{new_satscan_table}
\title{specific constructor for internal use}
\usage{
new_satscan_table(data, kind, spec = list())
}
\description{
specific constructor for internal use
}
\keyword{internal}
