% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{new_ss_tbl}
\alias{new_ss_tbl}
\title{Create a SaTScan Table}
\usage{
new_ss_tbl(data, type, roles, spec = list())
}
\arguments{
\item{data}{A data.frame.}

\item{type}{type string (e.g. "cas", "geo").}

\item{roles}{Named character vector mapping schema roles to data columns.}

\item{spec}{List of additional specifications.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
