% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NMlung_cas}
\alias{NMlung_cas}
\title{New Mexico Lung Cancer Data (Case File)}
\format{
A data frame with 7,793 rows and 5 variables:
\describe{
\item{county}{County name (Location ID).}
\item{cases}{Number of cases (1 for individual records).}
\item{date}{Date of diagnosis (Month/Year precision, day set to 1).}
\item{age_group}{Age group (18 levels: <5, 5-9, ..., 85+).}
\item{sex}{Sex (Male, Female).}
}
}
\source{
\url{https://www.satscan.org/datasets/nmlung/index.html}
}
\usage{
NMlung_cas
}
\description{
Lung cancer cases in New Mexico counties, 1973-1991.
}
\keyword{datasets}
