% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NMlung_geo}
\alias{NMlung_geo}
\title{New Mexico Lung Cancer Data (Coordinates)}
\format{
A data frame with 32 rows and 3 variables:
\describe{
\item{county}{County name (Location ID).}
\item{x_km}{X coordinate in km.}
\item{y_km}{Y coordinate in km.}
}
}
\source{
\url{https://www.satscan.org/datasets/nmlung/index.html}
}
\usage{
NMlung_geo
}
\description{
Centroid coordinates for New Mexico counties (Cartesian).
}
\keyword{datasets}
