% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NMlung_pop}
\alias{NMlung_pop}
\title{New Mexico Lung Cancer Data (Population File)}
\format{
A data frame with 66,115 rows and 5 variables:
\describe{
\item{county}{County name (Location ID).}
\item{year}{Census year (Integer).}
\item{population}{Population count.}
\item{age_group}{Age group (18 levels: <5, 5-9, ..., 85+).}
\item{sex}{Sex (Male, Female).}
}
}
\source{
\url{https://www.satscan.org/datasets/nmlung/index.html}
}
\usage{
NMlung_pop
}
\description{
Yearly population counts for New Mexico counties, 1973-1991.
}
\keyword{datasets}
