% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_results.R
\name{parse_results}
\alias{parse_results}
\title{Parse SatScan Results}
\usage{
parse_results(ss_results, data, id_quo, verbose = FALSE)
}
\arguments{
\item{ss_results}{Result from run_satscan()}

\item{data}{Original input data}

\item{id_quo}{Quosure for ID column}

\item{verbose}{Print debug info?}
}
\value{
data.frame with CLUSTER, P_VALUE, REL_RISK columns added
}
\description{
Extracts cluster info from rsatscan results and joins to original data
}
\keyword{internal}
