% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters.R
\name{prepare_cluster_data}
\alias{prepare_cluster_data}
\title{Prepare Cluster Data for Mapping}
\usage{
prepare_cluster_data(x, significance_only, coords)
}
\arguments{
\item{x}{satscan_result object}

\item{significance_only}{Filter for p < 0.05?}

\item{coords}{List with lat and lon column names}
}
\value{
Data frame ready for mapping
}
\description{
Prepare Cluster Data for Mapping
}
\keyword{internal}
