% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters.R
\name{prepare_location_data}
\alias{prepare_location_data}
\title{Prepare Location Data for Mapping}
\usage{
prepare_location_data(x, coords, popup_vars)
}
\arguments{
\item{x}{satscan_result object}

\item{coords}{List with lat and lon column names}

\item{popup_vars}{Additional variables for popups}
}
\value{
Data frame ready for mapping
}
\description{
Prepare Location Data for Mapping
}
\keyword{internal}
