% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_options.R
\name{print.satscan_prm}
\alias{print.satscan_prm}
\title{Print SaTScan Parameters}
\usage{
\method{print}{satscan_prm}(x, ...)
}
\arguments{
\item{x}{A \code{satscan_prm} object.}

\item{...}{Unused.}
}
\description{
Print SaTScan Parameters
}
