% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_result_s3_methods.R
\name{print.summary.satscan_result}
\alias{print.summary.satscan_result}
\title{Print Method for Summary of SaTScan Results}
\usage{
\method{print}{summary.satscan_result}(x, ...)
}
\arguments{
\item{x}{A \code{summary.satscan_result} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Method for Summary of SaTScan Results
}
