% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscanr.R
\name{print_prm_summary}
\alias{print_prm_summary}
\title{Print Parameter Summary}
\usage{
print_prm_summary(s)
}
\arguments{
\item{param_summary}{List from prm_summarize}
}
\description{
Print Parameter Summary
}
\keyword{internal}
