% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{print_satscan_summary}
\alias{print_satscan_summary}
\title{Print SatScan Parameter Summary}
\usage{
print_satscan_summary(params)
}
\arguments{
\item{params}{List. Parameter list}
}
\value{
Invisible NULL (prints to console)
}
\description{
Displays a formatted summary of SatScan parameters for review.
}
