% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{read_satscan_prm}
\alias{read_satscan_prm}
\title{Read SatScan Parameter File}
\usage{
read_satscan_prm(prm_path)
}
\arguments{
\item{prm_path}{Character. Path to the .prm file}
}
\value{
Named list of parameters
}
\description{
Reads a .prm file and returns a named list of parameters.
}
