% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_results.R
\name{run_satscan}
\alias{run_satscan}
\title{Run SatScan Analysis}
\usage{
run_satscan(
  work_dir,
  project_name = "epid",
  ss_location,
  ss_batch,
  verbose = FALSE
)
}
\arguments{
\item{work_dir}{Working directory with input files}

\item{project_name}{Project name (default: "epid")}

\item{ss_location}{SatScan install location}

\item{ss_batch}{SatScan batch file name}

\item{verbose}{Print output?}

\item{final_opts}{Options list for ss.options()}
}
\value{
satscan result object
}
\description{
Executes SatScan via rsatscan::satscan()
}
\keyword{internal}
