% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_parameters.R
\docType{data}
\name{satscan_parameters}
\alias{satscan_parameters}
\title{SaTScan Parameter Reference}
\description{
A comprehensive reference for all SaTScan parameters available via \code{epidscan}.
Use these parameters in a \code{.prm} template file or as overrides in \code{satscanr(...)}.
}
\details{
SaTScan is a software for spatial, temporal, and space-time scan statistics.
This documentation object serves as a placeholder for the parameter reference.

For a full, searchable list of all parameters, please see the package vignette:
\code{vignette("satscan-parameters", package = "epidscan")}

or visit the \href{http://www.satscan.org/}{SaTScan website} for full methodology.
}
\keyword{internal}
