% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_satscan_runner.R
\name{save_satscan_results}
\alias{save_satscan_results}
\title{Save SatScan Results}
\usage{
save_satscan_results(results, output_dir, project_name)
}
\arguments{
\item{results}{SatScan results object}

\item{output_dir}{Character. Directory to save results}

\item{project_name}{Character. Project name for output files}
}
\value{
List of paths to saved files
}
\description{
Saves SatScan results object and extracts cluster information to files.
}
