% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{set_satscan_path}
\alias{set_satscan_path}
\title{Set SatScan Executable Path}
\usage{
set_satscan_path(path)
}
\arguments{
\item{path}{Character. Absolute path to SatScan executable.}
}
\description{
Sets the path to the SatScan executable for the current session.
}
\examples{
\dontrun{
set_satscan_path("/Applications/SaTScan.app/Contents/app/satscan")
}
}
