% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_spatial_analysis.R
\name{smooth_raster}
\alias{smooth_raster}
\title{Smooth Population Raster}
\usage{
smooth_raster(pop_raster, sigma)
}
\arguments{
\item{pop_raster}{Input population raster}

\item{sigma}{Smoothing bandwidth}
}
\value{
Smoothed raster
}
\description{
Apply Gaussian smoothing to population raster to match the scale of analysis
}
