% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{ss_nbr}
\alias{ss_nbr}
\title{Coerce Data to SaTScan Neighbors Table}
\usage{
ss_nbr(data, loc_id, neighbor_cols)
}
\arguments{
\item{data}{A data frame containing neighbor lists.}

\item{loc_id}{Column name for the location ID.}

\item{neighbor_cols}{Character vector of column names representing neighbors.}
}
\value{
An \code{ss_tbl} object of type "nbr".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "nbr" (Neighbors File).
}
\section{SaTScan File Specification}{

The Neighbors File has the following structure:
\code{<LocationID> <Neighbor1ID> <Neighbor2ID> ...}

\itemize{
\item \strong{LocationID}: Identifier for the location.
\item \strong{NeighborIDs}: One or more columns, each representing a neighbor of the \code{LocationID}.
The order of these columns matters for SaTScan's interpretation of neighbor hierarchy.
}
}

