% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{ss_order}
\alias{ss_order}
\title{Get ss_tbl Column Order}
\usage{
ss_order(x)
}
\arguments{
\item{x}{An \code{ss_tbl} object.}
}
\value{
Character vector of column names in SaTScan output order.
}
\description{
Get ss_tbl Column Order
}
