% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{ss_pop}
\alias{ss_pop}
\title{Coerce Data to SaTScan Population Table}
\usage{
ss_pop(data, loc_id, time, population, covars = NULL, time_precision = NULL)
}
\arguments{
\item{data}{A data frame containing population data.}

\item{loc_id}{Column name for location ID.}

\item{time}{Column name for time (census year/time).}

\item{population}{Column name for population count.}

\item{covars}{Optional vector of column names for covariates.}

\item{time_precision}{User-specified time precision (e.g. "year" to format Dates as YYYY).}
}
\value{
An \code{ss_tbl} object of type "pop".
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Creates an \code{ss_tbl} object of type "pop" (Population File).
}
\section{SaTScan File Specification}{

The Population File has the following structure:
\code{<LocationID> <Year/Time> <Population> <Covariate1> ...}

\itemize{
\item \strong{LocationID}: Character or numeric identifier. Match cases and geo.
\item \strong{Year/Time}: The "census year" or time point. For daily analysis, you can still use
yearly population anchors (e.g. 2023, 2024). SaTScan interpolates linearly.
\item \strong{Population}: The count of people at risk.
\item \strong{Covariates}: Optional. If used, population must be stratified by these covariates.
}
}

