% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{ss_roles}
\alias{ss_roles}
\title{Get ss_tbl Roles}
\usage{
ss_roles(x)
}
\arguments{
\item{x}{An \code{ss_tbl} object.}
}
\value{
Named character vector of role mappings.
}
\description{
Get ss_tbl Roles
}
