% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prm_options.R
\name{summary.satscan_prm}
\alias{summary.satscan_prm}
\title{Summary of SaTScan Parameters}
\usage{
\method{summary}{satscan_prm}(object, ...)
}
\arguments{
\item{object}{A \code{satscan_prm} object.}

\item{...}{Unused.}
}
\description{
Summary of SaTScan Parameters
}
