% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_geocoding.R
\name{validate_coordinates}
\alias{validate_coordinates}
\title{Validate Coordinates}
\usage{
validate_coordinates(data, lat_col = "lat", long_col = "long")
}
\arguments{
\item{data}{Dataframe with lat/long columns}

\item{lat_col}{Name of latitude column}

\item{long_col}{Name of longitude column}
}
\value{
Dataframe with 'valid_geo' boolean column
}
\description{
Checks if coordinates fall within the expected bounding box (Songkhla).
}
