% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_validate_population_data.R
\name{validate_population_data}
\alias{validate_population_data}
\title{Validate Population Data}
\usage{
validate_population_data(data, mode = "simple")
}
\arguments{
\item{data}{Population data to validate}

\item{mode}{"simple" or "stratified"}
}
\value{
Validated data (throws error if invalid)
}
\description{
Validates population data structure and values before writing.
}
\examples{
validate_population_data(pop_data, mode = "simple")

}
