% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{write_satscan}
\alias{write_satscan}
\title{Write SaTScan File}
\usage{
write_satscan(x, file, sep = " ", na = "", quote = FALSE)
}
\arguments{
\item{x}{An ss_tbl object.}

\item{file}{Output file path.}

\item{sep}{Separator (default space).}

\item{na}{NA string (default "").}

\item{quote}{Logical (default FALSE).}
}
\description{
Writes an ss_tbl to disk in SaTScan ASCII format.
Handles column ordering, time formatting, and whitespace.
}
