% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_config.R
\name{write_satscan_params}
\alias{write_satscan_params}
\title{Write SatScan Parameters to File}
\usage{
write_satscan_params(params, output_dir, project_name)
}
\arguments{
\item{params}{List. Parameter list created by create_satscan_params()}

\item{output_dir}{Character. Directory where .prm file will be written}

\item{project_name}{Character. Project name (without extension)}
}
\value{
Path to created .prm file
}
\description{
Writes parameter list to .prm file using rsatscan.
}
\examples{
params <- create_satscan_params(start_date = "2025/01/01", end_date = "2025/11/30")
write_satscan_params(params, tempdir(), "lepto")
}
