% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_write_satscan_pop.R
\name{write_satscan_pop}
\alias{write_satscan_pop}
\title{Write SatScan Population File}
\usage{
write_satscan_pop(data, output_path, mode = "simple")
}
\arguments{
\item{data}{Processed population data}

\item{output_path}{Path to output file}

\item{mode}{"simple" or "stratified"}
}
\value{
Output file path (invisibly)
}
\description{
Writes population data to SatScan .pop file format.
}
\details{
File formats per SatScan manual:
- Simple: location_id, year, population
- Stratified: location_id, year, population, age_group, sex
}
\examples{
write_satscan_pop(pop_data, "output/satscan.pop", mode = "simple")

}
