% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_ss.R
\name{write_ss_file_wrapper}
\alias{write_ss_file_wrapper}
\title{Internal Helper: Write SS File Polymorphically}
\usage{
write_ss_file_wrapper(obj, path)
}
\description{
Internal Helper: Write SS File Polymorphically
}
\keyword{internal}
