% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_clusters.R
\name{create_cluster_popup}
\alias{create_cluster_popup}
\title{Create Cluster Popup HTML}
\usage{
create_cluster_popup(clusters)
}
\arguments{
\item{clusters}{Cluster data frame}
}
\value{
Character vector of HTML popup text
}
\description{
Create Cluster Popup HTML
}
\keyword{internal}
