% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_tbl.R
\name{new_ss_tbl}
\alias{new_ss_tbl}
\title{Create a SaTScan Table}
\usage{
new_ss_tbl(data, type, roles, spec = list())
}
\arguments{
\item{data}{A data.frame.}

\item{type}{type string (e.g. "cas", "geo").}

\item{roles}{Named character vector mapping schema roles to data columns.}

\item{spec}{List of additional specifications.}
}
\description{
Creates a new \code{ss_tbl} object, which is a \code{vctrs} subclass of \code{data.frame} tailored for SaTScan data.
}
