% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_results.R
\name{parse_satscan_output}
\alias{parse_satscan_output}
\title{Parse SaTScan Output into S3 Object}
\usage{
parse_satscan_output(
  ss_results,
  geo_df,
  loc_id_col,
  keep_raw = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{ss_results}{Result from read_satscan_files()}

\item{geo_df}{Data frame from ss_geo() with location IDs and coordinates}

\item{loc_id_col}{Name of the location ID column in geo_df}

\item{keep_raw}{Logical. Include raw output files?}

\item{verbose}{Print debug info?}
}
\value{
Object of class "satscan_result"
}
\description{
Parse SaTScan Output into S3 Object
}
\keyword{internal}
