% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/satscan_results.R
\name{run_satscan}
\alias{run_satscan}
\title{Run SatScan Analysis}
\usage{
run_satscan(
  work_dir,
  project_name = "epid",
  ss_location,
  ss_batch,
  verbose = FALSE
)
}
\arguments{
\item{work_dir}{Working directory with input files.}

\item{project_name}{Project name (default: "epid"). This matches the inputs CaseFile=epid.cas etc.}

\item{ss_location}{Path to the folder containing the SaTScan executable (Backwards compatibility param).}

\item{ss_batch}{Name of the SaTScan executable (Backwards compatibility param).}

\item{verbose}{Print output?}
}
\value{
satscan result object
}
\description{
Executes SatScan by calling the binary directly and parsing the output files.
Use \code{satscanr()} as the main entry point.
}
\keyword{internal}
